Rename version in x/staking legacy
