+++ x/staking/legacy/v0_38/types.go
@@ -1,6 +1,6 @@
 // DONTCOVER
 // nolint
-package v0_37
+package v0_38
 
 import (
 	"time"
